LPS-116954 portal-search-tuning-gsearch-configuration-web: Rename from es.js
