+++ modules/dxp/apps/portal-search-tuning/portal-search-tuning-gsearch-configuration-web/src/main/resources/META-INF/resources/edit_search_configuration.jsp
@@ -81,6 +81,6 @@ String redirect = ParamUtil.getString(request, "redirect", currentURL);
 
 	<react:component
 		data="<%= data %>"
-		module="js/ConfigurationSetApp.es"
+		module="js/ConfigurationSetApp"
 	/>
 </aui:form>
