+++ modules/dxp/apps/portal-search-tuning/portal-search-tuning-gsearch-configuration-web/src/main/resources/META-INF/resources/js/ConfigurationSetApp.js
@@ -11,7 +11,7 @@
 
 import React from 'react';
 
-import ConfigurationSetForm from './components/ConfigurationSetForm.es';
+import ConfigurationSetForm from './components/ConfigurationSetForm';
 
 export default function ({props}) {
 	return (
