+++ modules/dxp/apps/portal-search-tuning/portal-search-tuning-gsearch-configuration-web/src/main/resources/META-INF/resources/js/components/PageToolbar.js
@@ -16,7 +16,7 @@ import ClayManagementToolbar from '@clayui/management-toolbar';
 import PropTypes from 'prop-types';
 import React, {useContext, useState} from 'react';
 
-import ThemeContext from '../ThemeContext.es';
+import ThemeContext from '../ThemeContext';
 
 export default function PageToolbar({
 	onCancel,
